AG-17862 Extract dev-server markdown negotiation to shared factory
Move the SE-80 Accept: text/markdown dev-server negotiation mechanism into a
product-agnostic factory in ag-website-shared, parameterised by path patterns.
Grid's plugin becomes a thin wrapper supplying its own URL shapes; the charts
port reuses the factory. astro.config.mjs is unchanged.
